# packages/backend/openapi/openapi.yaml
...
openapi: '3.1.0'
info:
  title: Linked Data Explorer Backend
  description: >-
    DMN discovery and chain execution, SHACL validation, asset storage and
    integrations for the Linked Data Explorer.

    Every 4xx/5xx response is `application/problem+json` (RFC 9457,
    schema `Problem`): `status`, `title` and `detail` always; `code`, the
    machine-readable identifier existing callers branch on, where the
    handler sets one. Two operations are deliberate exceptions, each noted
    at its own response: `POST /dmns/evaluate/{decisionKey}` is a byte-for-byte
    proxy for Operaton's own evaluate endpoint and passes Operaton's error
    body through unchanged (or answers `OperatonProxyError` when Operaton
    could not be reached at all), and `GET /health`'s 503 body is the health
    document itself, not a problem response, because the document *is* the
    explanation. https://github.com/sgort/linked-data-explorer/issues/131.
  contact:
    name: Steven Gort
    email: steven.gort@ictu.nl
    url: https://iou-architectuur.open-regels.nl/
servers:
  - url: https://backend.linkeddata.open-regels.nl/v1
    description: Production
  - url: https://acc.backend.linkeddata.open-regels.nl/v1
    description: Acceptance
...
paths:
...
  /dmns:
    get:
      operationId: listDmns
      tags: [Discovery]
      summary: DMN decision models published in the knowledge graph
      description: >-
        Results are cached for five minutes per endpoint. The `endpoint` override
        is not validated (#142).
      parameters:
        - name: endpoint
          in: query
          required: false
          description: SPARQL endpoint URL to query instead of the configured one.
          schema:
            type: string
        - name: refresh
          in: query
          required: false
          description: '`true` or `1` bypasses the cache; any other value leaves it in use.'
          schema:
            type: string
      responses:
        '200':
          description: The decision models.
          headers:
            API-Version:
              $ref: '#/components/headers/ApiVersion'
          content:
            application/json:
              schema:
                type: object
                required: [success, data, timestamp]
                properties:
                  success:
                    const: true
                  data:
                    type: object
                    required: [total, dmns, fromCache]
                    properties:
                      total:
                        type: integer
                        minimum: 0
                      dmns:
                        type: array
                        items:
                          $ref: '#/components/schemas/DmnModel'
                      fromCache:
                        type: boolean
                        description: True unless `refresh` was requested; not a cache-hit indicator.
                  timestamp:
                    type: string
                    format: date-time
        '500':
          description: The knowledge graph could not be queried (code QUERY_ERROR).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /dmns/semantic-equivalences:
    get:
      operationId: listDmnSemanticEquivalences
      tags: [Discovery]
      summary: Variables of different DMNs linked to the same concept
      description: The `endpoint` override is not validated (#142).
      parameters:
        - name: endpoint
          in: query
          required: false
          description: SPARQL endpoint URL to query instead of the configured one.
          schema:
            type: string
      responses:
        '200':
          description: The equivalences.
          headers:
            API-Version:
              $ref: '#/components/headers/ApiVersion'
          content:
            application/json:
              schema:
                type: object
                required: [success, data, timestamp]
                properties:
                  success:
                    const: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SemanticEquivalence'
                  timestamp:
                    type: string
                    format: date-time
        '500':
          description: The knowledge graph could not be queried (code QUERY_ERROR).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /dmns/enhanced-chain-links:
    get:
      operationId: listDmnChainLinks
      tags: [Discovery]
      summary: Output-to-input links between DMNs, by identifier or shared concept
      description: The `endpoint` override is not validated (#142).
      parameters:
        - name: endpoint
          in: query
          required: false
          description: SPARQL endpoint URL to query instead of the configured one.
          schema:
            type: string
      responses:
        '200':
          description: The chain links.
          headers:
            API-Version:
              $ref: '#/components/headers/ApiVersion'
          content:
            application/json:
              schema:
                type: object
                required: [success, data, timestamp]
                properties:
                  success:
                    const: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EnhancedChainLink'
                  timestamp:
                    type: string
                    format: date-time
        '500':
          description: The knowledge graph could not be queried (code QUERY_ERROR).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /dmns/cycles:
    get:
      operationId: listDmnChainCycles
      tags: [Discovery]
      summary: Cycles in the DMN chain graph
      description: The `endpoint` override is not validated (#142).
      parameters:
        - name: endpoint
          in: query
          required: false
          description: SPARQL endpoint URL to query instead of the configured one.
          schema:
            type: string
      responses:
        '200':
          description: The detected cycles.
          headers:
            API-Version:
              $ref: '#/components/headers/ApiVersion'
          content:
            application/json:
              schema:
                type: object
                required: [success, data, timestamp]
                properties:
                  success:
                    const: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ChainCycle'
                  timestamp:
                    type: string
                    format: date-time
        '500':
          description: The knowledge graph could not be queried (code QUERY_ERROR).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /dmns/{identifier}:
    get:
      operationId: getDmn
      tags: [Discovery]
      summary: One DMN decision model
      description: Served from the same five-minute cache as the list. The `endpoint` override is not validated (#142).
      parameters:
        - name: identifier
          in: path
          required: true
          description: The DMN's identifier.
          schema:
            type: string
        - name: endpoint
          in: query
          required: false
          description: SPARQL endpoint URL to query instead of the configured one.
          schema:
            type: string
      responses:
        '200':
          description: The decision model.
          headers:
            API-Version:
              $ref: '#/components/headers/ApiVersion'
          content:
            application/json:
              schema:
                type: object
                required: [success, data, timestamp]
                properties:
                  success:
                    const: true
                  data:
                    $ref: '#/components/schemas/DmnModel'
                  timestamp:
                    type: string
                    format: date-time
        '404':
          description: No DMN has this identifier (code NOT_FOUND).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: The knowledge graph could not be queried (code QUERY_ERROR).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /dmns/{identifier}/xml:
    get:
      operationId: getDmnXml
      tags: [Discovery]
      summary: The deployed DMN XML, as a file download
      description: Fetched from Operaton; `identifier` is used as the decision definition key.
      parameters:
        - name: identifier
          in: path
          required: true
          description: The DMN's identifier, used as the Operaton decision definition key.
          schema:
            type: string
      responses:
        '200':
          description: The DMN XML.
          headers:
            API-Version:
              $ref: '#/components/headers/ApiVersion'
            Content-Disposition:
              description: '`attachment; filename="<identifier>.dmn"`.'
              schema:
                type: string
          content:
            application/xml:
              schema:
                type: string
        '404':
          description: Operaton has no decision definition with this key (code DMN_NOT_FOUND).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Operaton could not be reached or answered with an error (code DMN_FETCH_FAILED).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /dmns/drd/deploy:
    post:
      operationId: deployDrd
      tags: [Discovery]
      summary: Assemble deployed DMNs into one DRD and deploy it to Operaton
      description: >-
        The last id in `dmnIds` is the entry point and names the deployed file.
        Operaton's own error status is not passed through; failures answer 500.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [dmnIds, deploymentName]
              properties:
                dmnIds:
                  type: array
                  minItems: 2
                  items:
                    type: string
                deploymentName:
                  type: string
      responses:
        '200':
          description: The DRD was deployed.
          headers:
            API-Version:
              $ref: '#/components/headers/ApiVersion'
          content:
            application/json:
              schema:
                type: object
                required: [success, data, timestamp]
                properties:
                  success:
                    const: true
                  data:
                    type: object
                    required: [deploymentId, entryPointId, filename, dmnCount]
                    properties:
                      deploymentId:
                        type: string
                      entryPointId:
                        type: string
                      filename:
                        type: string
                      dmnCount:
                        type: integer
                        minimum: 2
                  timestamp:
                    type: string
                    format: date-time
        '400':
          description: '`dmnIds` has fewer than two entries or `deploymentName` is blank (code INVALID_INPUT).'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: >-
            Assembly or deployment failed (code DRD_DEPLOY_FAILED), or the request
            body was not valid JSON (code INTERNAL_ERROR; #143).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
  /dmns/deploy:
    post:
      operationId: deployDmn
      tags: [Discovery]
      summary: Deploy DMN XML to Operaton as it is
      description: Operaton's own error status is not passed through; failures answer 500.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [xml, deploymentName]
              properties:
                xml:
                  type: string
                deploymentName:
                  type: string
                filename:
                  type: string
                  description: Defaults to `<deploymentName>.dmn` when missing or blank.
      responses:
        '200':
          description: The DMN was deployed.
          headers:
            API-Version:
              $ref: '#/components/headers/ApiVersion'
          content:
            application/json:
              schema:
                type: object
                required: [success, data, timestamp]
                properties:
                  success:
                    const: true
                  data:
                    type: object
                    required: [deploymentId]
                    properties:
                      deploymentId:
                        type: string
                  timestamp:
                    type: string
                    format: date-time
        '400':
          description: '`xml` or `deploymentName` is missing or blank (code INVALID_INPUT).'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: >-
            Deployment failed (code DMN_DEPLOY_FAILED), or the request body was not
            valid JSON (code INTERNAL_ERROR; #143).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
components:
  headers:
    ApiVersion:
      description: Full release version of the API (/core/version-header).
      required: true
      schema:
        type: string
        examples:
          - '2026.09.4'
  schemas:
...
    SemanticEquivalence:
      type: object
      required: [sharedConcept, concept1, concept2, dmn1, dmn2]
      properties:
        sharedConcept:
          type: string
          description: IRI of the concept both variables refer to.
        concept1:
          $ref: '#/components/schemas/SemanticConcept'
        concept2:
          $ref: '#/components/schemas/SemanticConcept'
        dmn1:
          $ref: '#/components/schemas/DmnReference'
        dmn2:
          $ref: '#/components/schemas/DmnReference'
...
    DmnReference:
      type: object
      required: [uri, title]
      properties:
        uri:
          type: string
        identifier:
          type: string
        title:
          type: string
...
    ChainCycle:
      type: object
      required: [path, type]
      properties:
        path:
          type: array
          minItems: 3
          items:
            $ref: '#/components/schemas/DmnReference'
        type:
          type: string
          enum: [three-hop, four-hop]
          description: Only `three-hop` cycles are detected today.
...